Fix problem with global indexer when a population is passed.
Problem happens if the population is a mix of two different populations.

Actually you should provide a helper function to mix two populations, and the helper function should take into account the global indexer. This will become more of a nuisance with sexual reproduction - with crossover, the node number does matter! For now it's fine though.